Refactor checker method and add file for class
